Refactor data science utils
