+++ zincbase/utils/data_science.py
@@ -1,6 +1,8 @@
 import csv
 import re
 
+from sklearn.metrics import roc_auc_score
+from sklearn.preprocessing import LabelBinarizer
 from tqdm import tqdm
 
 from utils.string_utils import cleanse
@@ -36,3 +38,16 @@ def calc_mrr(kb, test_file, delimiter=',', header=None, size=None):
         except:
             continue
     return mrr / size
+
+def calc_auc_roc(truth, pred, average="macro"):
+    """Calculate the Area-Under-the-Curve Receiver Operating Characteristic
+
+    A funny measure that combines precision and recall. Sklearn can't agree how
+    to implement it for multiclass; this version is from fbrundu on
+    https://github.com/scikit-learn/scikit-learn/issues/3298
+    """
+    lb = LabelBinarizer()
+    lb.fit(truth)
+    truth = lb.transform(truth)
+    pred = lb.transform(pred)
+    return roc_auc_score(truth, pred, average=average)
